+++ omega-surface/private/omega_vertsolve_codegen.m
@@ -35,7 +35,7 @@ function omega_vertsolve_codegen(nk, ni, nj, Xvec, OPTS)
 % Author(s) : Geoff Stanley
 % Email     : g.stanley@unsw.edu.au
 % Email     : geoffstanley@gmail.com
-% Version   : 2.1.0
+% Version   : 2.1.1
 %
 % Modified by : --
 % Date        : --
@@ -102,7 +102,7 @@ try
         else
             t_X  = coder.typeof(0, [nk, ni, nj], [vs, vs, vs]);
         end
-        t_BotK   = coder.typeof(uint16(0), [ni, nj], [vs, vs]);
+        t_BotK   = coder.typeof(0, [ni, nj], [vs, vs]);
         t_x      = coder.typeof(0, [ni, nj], [vs, vs]);
         
         args = {t_SppX, t_SppX, t_X, t_BotK, t_x, t_x, t_x, 0, t_x};
